move leaderboard to its own app
